AIClient/requirements.txt
websockets
python-dotenv
llama-cpp-python
